New Foundation
Changed the structure of Player Boxes from simple layout collections
into Fragments for modularization.  This results in similar behaviors
and looks to the user, but significantly easier to update code for new
features later on.  Also added a new feature, the LongPress on minus or
plus button, as an example of how creating these becomes easier by
fragments.  On a LongPress of a button, the health value now decrements
/ increments by 5, rather than 1 for a normal press.