Add and Remove Players
Added action bar buttons to add and remove players in the game.
Players are now limited to four maximum and one minimum, due to size
constraints on most phones, and logically a game cannot exist without
at least one player.  The maximum player limit might eventually change
depending on the screen size of the user, or the structure of a player
box.